+++ include/Arcade/Game/IGameModule.hpp
@@ -34,7 +34,7 @@ namespace Arcade {
                  * @param eventManager the event manager
                  */
                 virtual void update(
-                float deltaTime, Arcade::ECS::IEventManager &eventManager) = 0;
+                double deltaTime, Arcade::ECS::IEventManager &eventManager) = 0;
                 /**
                  * @brief Get the entity manager of the current active scene
                  *
